django_site/requirements.txt
django>=3.2,<3.3
django-countries==7.1
django-ratelimit
